+++ core/core-api/src/main/java/com/ticket/core/api/controller/HoldController.java
@@ -3,7 +3,7 @@ package com.ticket.core.api.controller;
 import com.ticket.core.api.controller.docs.HoldControllerDocs;
 import com.ticket.core.api.controller.request.CreateHoldRequest;
 import com.ticket.core.domain.member.MemberPrincipal;
-import com.ticket.core.domain.order.command.usecase.StartOrderUseCase;
+import com.ticket.core.domain.order.create.CreateOrderUseCase;
 import com.ticket.core.support.response.ApiResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -22,18 +22,18 @@ import java.net.URI;
 @RequiredArgsConstructor
 public class HoldController implements HoldControllerDocs {
 
-    private final StartOrderUseCase startOrderUseCase;
+    private final CreateOrderUseCase createOrderUseCase;
 
     @Override
     @PostMapping
-    public ResponseEntity<ApiResponse<StartOrderUseCase.Output>> createHold(
+    public ResponseEntity<ApiResponse<CreateOrderUseCase.Output>> createHold(
             @PathVariable final Long performanceId,
             @Valid @RequestBody final CreateHoldRequest request,
             final MemberPrincipal memberPrincipal
     ) {
-        final StartOrderUseCase.Input input =
-                new StartOrderUseCase.Input(performanceId, request.getSeatIds(), memberPrincipal.getMemberId());
-        final StartOrderUseCase.Output output = startOrderUseCase.execute(input);
+        final CreateOrderUseCase.Input input =
+                new CreateOrderUseCase.Input(performanceId, request.getSeatIds(), memberPrincipal.getMemberId());
+        final CreateOrderUseCase.Output output = createOrderUseCase.execute(input);
         return ResponseEntity.created(URI.create("/api/v1/orders/" + output.orderKey()))
                 .header("X-Order-Key", output.orderKey())
                 .body(ApiResponse.success(output));
