+++ core/core-api/src/main/java/com/ticket/core/api/controller/docs/OrderControllerDocs.java
@@ -1,9 +1,9 @@
 package com.ticket.core.api.controller.docs;
 
-import com.ticket.core.api.controller.request.StartOrderRequest;
+import com.ticket.core.api.controller.request.CreateOrderRequest;
 import com.ticket.core.domain.member.MemberPrincipal;
-import com.ticket.core.domain.order.command.usecase.StartOrderUseCase;
-import com.ticket.core.domain.order.query.usecase.GetOrderDetailUseCase;
+import com.ticket.core.domain.order.create.CreateOrderUseCase;
+import com.ticket.core.domain.order.query.GetOrderDetailUseCase;
 import com.ticket.core.support.response.ApiResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -20,8 +20,8 @@ public interface OrderControllerDocs {
     @Operation(
             summary = "주문 시작",
             description = """
-                    요청한 좌석을 선점하고 결제 진입용 PENDING 주문을 시작합니다.
-                    동일 회원은 같은 회차에 PENDING 주문을 1건만 가질 수 있습니다.
+                    요청한 좌석을 선점하고 결제 진입용 PENDING 주문을 생성합니다.
+                    동일 회원과 같은 공연에는 PENDING 주문을 1건만 가질 수 있습니다.
                     응답 헤더로 생성된 주문 조회 URI(Location)와 주문 키(X-Order-Key)를 반환합니다.
                     생성된 주문은 GET /api/v1/orders/{orderKey}로 조회할 수 있습니다.
                     """
@@ -45,16 +45,16 @@ public interface OrderControllerDocs {
             ),
             @io.swagger.v3.oas.annotations.responses.ApiResponse(
                     responseCode = "409",
-                    description = "이미 선점된 좌석이 있거나 진행 중인 PENDING 주문이 존재함",
+                    description = "이미 선점된 좌석이 있거나 진행 중인 PENDING 주문이 존재",
                     content = @Content(schema = @Schema(implementation = com.ticket.core.support.response.ApiResponse.class))
             )
     })
-    ResponseEntity<ApiResponse<StartOrderUseCase.Output>> startOrder(
-            StartOrderRequest request,
+    ResponseEntity<ApiResponse<CreateOrderUseCase.Output>> createOrder(
+            CreateOrderRequest request,
             @Parameter(hidden = true) MemberPrincipal memberPrincipal
     );
 
-    @Operation(summary = "주문 조회", description = "주문/결제 화면 진입 때 호출합니다. 만료된 주문은 EXPIRED 상태로 표시됩니다.")
+    @Operation(summary = "주문 조회", description = "주문/결제 화면 진입 시 호출합니다. 만료된 주문은 EXPIRED 상태로 표시됩니다.")
     @ApiResponses(value = {
             @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "200", description = "주문 조회 성공"),
             @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "403", description = "본인 주문이 아님")
