--- core/core-api/src/main/java/com/ticket/core/domain/order/application/OrderKeyGenerator.java
@@ -1,13 +1,0 @@
-package com.ticket.core.domain.order.application;
-
-import org.springframework.stereotype.Component;
-
-import java.util.UUID;
-
-@Component
-public class OrderKeyGenerator {
-
-    public String generate() {
-        return "ORDER-" + UUID.randomUUID().toString().replace("-", "");
-    }
-}
