--- core/core-api/src/main/java/com/ticket/core/domain/order/command/usecase/RequestedSeatIds.java
@@ -1,51 +1,0 @@
-package com.ticket.core.domain.order.command.usecase;
-
-import com.ticket.core.support.exception.CoreException;
-import com.ticket.core.support.exception.ErrorType;
-
-import java.util.HashSet;
-import java.util.List;
-
-public final class RequestedSeatIds {
-
-    private final List<Long> values;
-
-    private RequestedSeatIds(final List<Long> values) {
-        this.values = values;
-    }
-
-    public static RequestedSeatIds from(final List<Long> requestedSeatIds) {
-        validateDuplicate(requestedSeatIds);
-        return new RequestedSeatIds(normalize(requestedSeatIds));
-    }
-
-    private static void validateDuplicate(final List<Long> requestedSeatIds) {
-        if (requestedSeatIds.size() == new HashSet<>(requestedSeatIds).size()) {
-            return;
-        }
-        throw new CoreException(ErrorType.INVALID_REQUEST, "중복된 seatId가 포함되어 있습니다.");
-    }
-
-    private static List<Long> normalize(final List<Long> requestedSeatIds) {
-        final List<Long> seatIds = requestedSeatIds.stream()
-                .sorted()
-                .toList();
-        validateEmpty(seatIds);
-        return seatIds;
-    }
-
-    private static void validateEmpty(final List<Long> seatIds) {
-        if (!seatIds.isEmpty()) {
-            return;
-        }
-        throw new CoreException(ErrorType.INVALID_REQUEST, "선점할 좌석이 없습니다.");
-    }
-
-    public int size() {
-        return values.size();
-    }
-
-    public List<Long> values() {
-        return values;
-    }
-}
