+++ core/core-api/src/main/java/com/ticket/core/domain/order/create/CreateOrderUseCase.java
@@ -1,0 +1,71 @@
+package com.ticket.core.domain.order.create;
+
+import com.ticket.core.aop.DistributedLock;
+import com.ticket.core.domain.hold.application.HoldHistoryRecorder;
+import com.ticket.core.domain.hold.event.HoldCreatedEvent;
+import com.ticket.core.domain.order.model.Order;
+import com.ticket.core.domain.performance.Performance;
+import com.ticket.core.enums.OrderState;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.ApplicationEventPublisher;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.time.Duration;
+import java.time.LocalDateTime;
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+public class CreateOrderUseCase {
+
+    private final CreateOrderValidator validator;
+    private final HoldAllocator holdAllocator;
+    private final OrderCreator orderCreator;
+    private final HoldHistoryRecorder holdHistoryRecorder;
+    private final ApplicationEventPublisher applicationEventPublisher;
+
+    public record Input(Long performanceId, List<Long> seatIds, Long memberId) {}
+
+    public record Output(String orderKey, OrderState status, LocalDateTime expiresAt) {}
+
+    @Transactional
+    @DistributedLock(
+            prefix = "start-order",
+            dynamicKey = "#input.memberId() + ':' + #input.performanceId()",
+            leaseTime = 15000L,
+            message = "주문 시작 처리 중입니다. 잠시 후 다시 시도해 주세요."
+    )
+    public Output execute(final Input input) {
+        final RequestedSeatIds requestedSeatIds = RequestedSeatIds.from(input.seatIds());
+        final Performance performance = validator.validate(input.memberId(), input.performanceId(), requestedSeatIds);
+        final Duration holdDuration = Duration.ofSeconds(performance.getHoldTime());
+        final HoldAllocation allocation = holdAllocator.allocate(input.memberId(), input.performanceId(), requestedSeatIds, holdDuration);
+        try {
+            return createOrder(input, holdDuration, allocation);
+        } catch (final RuntimeException e) {
+            holdAllocator.release(allocation);
+            throw e;
+        }
+    }
+
+    private Output createOrder(final Input input, final Duration holdDuration, final HoldAllocation allocation) {
+        final Order order = orderCreator.createPendingOrder(
+                input.memberId(),
+                input.performanceId(),
+                allocation.holdKey(),
+                allocation.expiresAt(),
+                allocation.performanceSeats()
+        );
+        holdHistoryRecorder.recordCreated(
+                input.memberId(),
+                input.performanceId(),
+                allocation.holdKey(),
+                allocation.startedAt(holdDuration),
+                allocation.expiresAt(),
+                allocation.performanceSeats()
+        );
+        applicationEventPublisher.publishEvent(new HoldCreatedEvent(allocation.snapshot()));
+        return new Output(order.getOrderKey(), OrderState.PENDING, allocation.expiresAt());
+    }
+}
