+++ core/core-api/src/main/java/com/ticket/core/domain/order/create/OrderKeyGenerator.java
@@ -1,0 +1,13 @@
+package com.ticket.core.domain.order.create;
+
+import org.springframework.stereotype.Component;
+
+import java.util.UUID;
+
+@Component
+public class OrderKeyGenerator {
+
+    public String generate() {
+        return "ORDER-" + UUID.randomUUID().toString().replace("-", "");
+    }
+}
