+++ core/core-api/src/main/java/com/ticket/core/domain/order/shared/OrderTerminationResult.java
@@ -1,0 +1,10 @@
+package com.ticket.core.domain.order.shared;
+
+import java.util.List;
+
+public record OrderTerminationResult(
+        Long performanceId,
+        String holdKey,
+        List<Long> seatIds
+) {
+}
